pyproject.toml
[tool.poetry]
name = "cofense"
version = "0.0.1"
description = "OpenCTI connector for Cofense ThreatHQ"
authors = []

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
pycti = "5.5.6"
pydantic = "^1.9.1"
stix2-elevator = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"